Restriction fixed
Fixed the restriction on the CY metric, put the restriction of z to the left,
and the restriction on zbar to the right.
Also replace the delete_column() in front of the the partial derivatives by the
restriction matrices when calculating the one-forms. As the partial derivatives
should also be restricted